refactor: 클래스의 역할에 맞게 패키지를 변경하라
* config:
  - AppConfig.java
  - InputValidator.java
  - RacingInitializer.java
  - RacingConstants.java
* controller:
  - InputController.java
* service:
  - Car.java
  - CarRacing.java
  - Racers.java
  - RacingRule.java
* view:
  - InputView.java
  - OutputView.java
  - ViewConstants.java
