+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.config.AppConfig;
+import racingcar.service.CarRacing;
+
 public class Application {
     public static void main(String[] args) {
         AppConfig appConfig = new AppConfig();
