+++ src/main/java/racingcar/config/RacingInitializer.java
@@ -1,8 +1,12 @@
-package racingcar;
+package racingcar.config;
 
-import static racingcar.RacingConstants.INITIAL_MOVE_COUNT;
+import static racingcar.config.RacingConstants.INITIAL_MOVE_COUNT;
 
 import java.util.List;
+import racingcar.service.Car;
+import racingcar.controller.InputController;
+import racingcar.service.Racers;
+import racingcar.View.InputView;
 
 public class RacingInitializer {
     private final InputView inputView;
@@ -16,7 +20,9 @@ public class RacingInitializer {
     public Racers registerRacers() {
         String inputNames = inputView.requestCarNames();
         List<String> carNames = inputController.extractCarNames(inputNames);
-        return new Racers(carNames.stream().map(name -> new Car(name, INITIAL_MOVE_COUNT)).toList());
+        return new Racers(carNames.stream()
+                .map(name -> new Car(name, INITIAL_MOVE_COUNT))
+                .toList());
     }
 
     public int registerTotalRounds() {
