+++ src/main/java/racingcar/service/RacingRule.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
