+++ src/test/java/racingcar/CarTest.java
@@ -2,15 +2,16 @@ package racingcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static racingcar.RacingConstants.INITIAL_MOVE_COUNT;
-import static racingcar.RacingConstants.MAX_LENGTH_OF_CAR_NAME;
-import static racingcar.RacingConstants.MIN_LENGTH_OF_CAR_NAME;
+import static racingcar.config.RacingConstants.INITIAL_MOVE_COUNT;
+import static racingcar.config.RacingConstants.MAX_LENGTH_OF_CAR_NAME;
+import static racingcar.config.RacingConstants.MIN_LENGTH_OF_CAR_NAME;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.service.Car;
 
 class CarTest {
     @DisplayName("최소 길이보다 짧은 이름을 전달받으면 예외 발생")
