+++ src/test/java/racingcar/InputValidatorTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.config.InputValidator;
 
 class InputValidatorTest {
     @DisplayName("숫자로만 구성된 문자열이면 통과")
